// src/config/constants/types.ts
export type Address = {[chain in ChainId | string]: string}

export enum QuoteToken {
  'BNB' = 'BNB',
  'CAKE' = 'CAKE',
  'SYRUP' = 'SYRUP',
  'BUSD' = 'BUSD',
  'TWT' = 'TWT',
  'UST' = 'UST',
  'WCAKE' = 'WCAKE',
  'VAULT' = 'VAULT',
  'BTCB' = 'BTCB',
  'MASH' = 'MASH'
}
export interface SerializedToken {
  chainId: number
  address: string
  decimals: number
  symbol?: string
  name?: string
  projectLink?: string
  logoURI?: string
}
// ...
interface FarmConfigBaseProps {
  pid: number
  v1pid?: number
  lpSymbol: string
  defaultApr: string,
  lpAddresses: Address
  multiplier?: string
  isCommunity?: boolean
  depositFeeBP: number
  dual?: {
    rewardPerBlock: number
    earnLabel: string
    endBlock: number
  }
}

export interface SerializedFarmConfig extends FarmConfigBaseProps {
  token: SerializedToken
  quoteToken: SerializedToken
}